+++ server/routes.js
@@ -17,8 +17,7 @@ import { stripeWebhook, transferwiseWebhook } from './controllers/webhooks';
 import * as email from './controllers/services/email';
 
 import required from './middleware/required_param';
-import * as aN from './middleware/security/authentication';
-import * as auth from './middleware/security/auth';
+import * as authentication from './middleware/authentication';
 import errorHandler from './middleware/error_handler';
 import * as params from './middleware/params';
 import sanitizer from './middleware/sanitizer';
@@ -48,9 +47,9 @@ export default app => {
     next();
   });
 
-  app.use('*', auth.checkClientApp);
+  app.use('*', authentication.checkClientApp);
 
-  app.use('*', auth.authorizeClientApp);
+  app.use('*', authentication.authorizeClientApp);
 
   // Setup rate limiter
   if (get(config, 'redis.serverUrl')) {
@@ -95,13 +94,13 @@ export default app => {
    * User reset password or new token flow (no jwt verification)
    */
   app.post('/users/signin', required('user'), users.signin);
-  app.post('/users/update-token', auth.mustBeLoggedIn, users.updateToken);
+  app.post('/users/update-token', authentication.mustBeLoggedIn, users.updateToken);
 
   /**
    * Moving forward, all requests will try to authenticate the user if there is a JWT token provided
    * (an error will be returned if the JWT token is invalid, if not present it will simply continue)
    */
-  app.use('*', aN.authenticateUser); // populate req.remoteUser if JWT token provided in the request
+  app.use('*', authentication.authenticateUser); // populate req.remoteUser if JWT token provided in the request
 
   /**
    * Parameters.
@@ -157,8 +156,8 @@ export default app => {
   app.post('/webhooks/stripe', stripeWebhook); // when it gets a new subscription invoice
   app.post('/webhooks/transferwise', transferwiseWebhook); // when it gets a new subscription invoice
   app.post('/webhooks/mailgun', email.webhook); // when receiving an email
-  app.get('/connected-accounts/:service/callback', aN.authenticateServiceCallback); // oauth callback
-  app.delete('/connected-accounts/:service/disconnect/:collectiveId', aN.authenticateServiceDisconnect);
+  app.get('/connected-accounts/:service/callback', authentication.authenticateServiceCallback); // oauth callback
+  app.delete('/connected-accounts/:service/disconnect/:collectiveId', authentication.authenticateServiceDisconnect);
 
   app.use(sanitizer()); // note: this break /webhooks/mailgun /graphiql
 
@@ -182,9 +181,12 @@ export default app => {
   /**
    * Generic OAuth (ConnectedAccounts)
    */
-  app.get('/connected-accounts/:service(github)', aN.authenticateService); // backward compatibility
-  app.get('/connected-accounts/:service(github|twitter|meetup|stripe|paypal)/oauthUrl', aN.authenticateService);
-  app.get('/connected-accounts/:service/verify', aN.parseJwtNoExpiryCheck, connectedAccounts.verify);
+  app.get('/connected-accounts/:service(github)', authentication.authenticateService); // backward compatibility
+  app.get(
+    '/connected-accounts/:service(github|twitter|meetup|stripe|paypal)/oauthUrl',
+    authentication.authenticateService,
+  );
+  app.get('/connected-accounts/:service/verify', authentication.parseJwtNoExpiryCheck, connectedAccounts.verify);
 
   /* PayPal Payment Method Helpers */
   app.post('/services/paypal/create-payment', paypal.createPayment);
